// app/src/components/trace/__generated__/SpanAnnotationsEditorQuery.graphql.ts
export type SpanAnnotationsEditorQuery$variables = {
  spanId: string;
};
export type SpanAnnotationsEditorQuery$data = {
  readonly span: {
    readonly id: string;
    readonly " $fragmentSpreads": FragmentRefs<"SpanAnnotationsEditor_spanAnnotations">;
  };
};
export type SpanAnnotationsEditorQuery = {
  response: SpanAnnotationsEditorQuery$data;
  variables: SpanAnnotationsEditorQuery$variables;
};