// app/src/components/trace/__generated__/SpanAnnotationsEditorSpanAnnotationsQuery.graphql.ts
export type SpanAnnotationsEditorSpanAnnotationsQuery$variables = {
  id: string;
};
export type SpanAnnotationsEditorSpanAnnotationsQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"SpanAnnotationsEditor_spanAnnotations">;
  };
};
export type SpanAnnotationsEditorSpanAnnotationsQuery = {
  response: SpanAnnotationsEditorSpanAnnotationsQuery$data;
  variables: SpanAnnotationsEditorSpanAnnotationsQuery$variables;
};